benAnadolu_con/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.3
django-ckeditor==6.1.0
django-classy-tags==0.6.2
django-colorfield==0.4.2
django-js-asset==1.2.2
httplib2==0.19.1
idna==3.2
oauthlib==3.1.1
pafy==0.3.82
Pillow==8.2.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
